Update README.md
Olympic is the most popular international variety of sports competition which is held in Summer and Winter where a number of athletes over the world participate. The Modern Olympics competitions are held every four years as Summer and Winter cycles, with the Winter cycles held in 1994, 1998, and so on and the Summer cycles in the year 1996, 2000, and so on. There are two long periods 1912-1920 and 1936-1948 when Olympics competitions were skipped because of World War I and II.
As the Olympics are sports competitions, it is important to know which country was dominant or in simple language which country won a greater number of medals. Another question is what age group has won a greater number of medals. So, My Story focuses on analyzing the top 20 countries that won more medals overall. Also, I have analyzed the number of male and female participants that won the medals according to their age.
The outcome of the visualizations will be detailed information of yearly country-wise medal tally for Olympics and more clarity on the age group in males and females that won a greater number of medals.